allow reading device's IP address from channel
Dialplan example:

   NoOp(device IP: ${CHANNEL(peerip)})

Returns the IP address in dotted-quad format of the SCCP device linked
to the channel.

Similar as chan_sip "${CHANNEL(peerip)}".

Note that there is a little thread safety issue in the
sccp_channel_tech_acf_channel_read function, but the real problem comes
from the fact that Asterisk is calling this function without the channel
being locked (i.e. there's a thread safety issue in the function calling
this function), so it ain't worth fixing (since it will be only a
partial and ugly fix). In practice, this shouldn't cause any real
problem the way we intend to use it.