add "max_guests" option
The max_guests option limits the number of guest devices that can be
connected at any time.  Although you might want to turn off the guest
feature completely if you are in a hostile environment, the max_guest
option makes it much less worse to leave it enabled since an attacker
can't use the guest feature to exhaust all file descriptors (and other
resources) of your asterisk.

To make the option work properly and keep the implementation simple, the
"guest" state of a device can't change during the sccp_device lifetime.
Otherwise, the count of guest devices would need to be adjusted when the
device "guest" state change, and since for all practical purpose, a
guest <-> non-guest config switch triggers a reset/reconnection, having
additional code to handle that doesn't seem to be worth.

Also, if the max_guest option is decreased below the current number of
connected guest devices, no guest devices are forcefully disconnected,
to again simplify the implementation and because I felt like it was an
acceptable behaviour.

Note that calling a guest device is still someting not really supported
(since they all share the same line name, if you have more than 1 guest
device connected when you dial e.g. SCCP/guest, a random one will be called).
This would be something to review, I don't know if there's other
consequence to that.